asn_private: Initial copilot split into public and private modeles
