+++ romancal/associations/lib/association_rules.py
@@ -6,6 +6,6 @@ These associations are specifically defined for use in DMS.
 """
 
 from romancal.associations.lib import rules_level2
-from romancal.associations.registry import RegistryMarker
+from romancal.associations._registry import RegistryMarker
 
 RegistryMarker.mark(rules_level2)
