+++ romancal/associations/lib/dms_base.py
@@ -1,6 +1,6 @@
 """Association attributes common to DMS-based Rules"""
 
-from romancal.associations.exceptions import AssociationNotValidError
+from romancal.associations._exceptions import AssociationNotValidError
 from romancal.associations.lib.acid import ACIDMixin
 from romancal.associations.lib.constraint import (
     AttrConstraint,
