+++ romancal/associations/lib/rules_elpp_base.py
@@ -12,8 +12,8 @@ from typing import TYPE_CHECKING
 from stpipe.format_template import FormatTemplate
 
 from romancal.associations import libpath
-from romancal.associations.association import Association
-from romancal.associations.exceptions import AssociationNotValidError
+from romancal.associations._association import Association
+from romancal.associations._exceptions import AssociationNotValidError
 from romancal.associations.lib.acid import ACID
 from romancal.associations.lib.constraint import Constraint, SimpleConstraint
 from romancal.associations.lib.counter import Counter
@@ -34,7 +34,7 @@ from romancal.associations.lib.product_utils import (
     prune_duplicate_products,
 )
 from romancal.associations.lib.utilities import evaluate, is_iterable
-from romancal.associations.registry import RegistryMarker
+from romancal.associations._registry import RegistryMarker
 
 if TYPE_CHECKING:
     from typing import ClassVar
