+++ romancal/associations/tests/test_asn_from_list.py
@@ -4,7 +4,7 @@ import pytest
 
 from romancal.associations import Association, AssociationRegistry, load_asn
 from romancal.associations.asn_from_list import _cli, asn_from_list
-from romancal.associations.exceptions import AssociationNotValidError
+from romancal.associations._exceptions import AssociationNotValidError
 
 
 def test_base_association():
