+++ romancal/associations/tests/test_level2_basics.py
@@ -1,7 +1,7 @@
 """Test basic usage of Level2 associations"""
 
 from romancal.associations import generate, load_asn
-from romancal.associations.main import Main
+from romancal.associations._main import Main
 from romancal.associations.tests.helpers import (
     combine_pools,
     registry_level2_only,
